requirements.txt
flask
chromadb
beautifulsoup4
requests
google-generativeai
